Project setup, read symbols from ELF file
The project uses mcppbs (Modular C++ Build System), similarly as the
RISCV simulator and the proxy kernel. The project is described in
configure.ac, with additional info in the parser/ folder. The needed
scripts are in the scripts/ folder.

Implemented reading symbols from the ELF file. Currently only 64-bit
ELF files are supported. The data is stored in an object of class
elf_data_t.