Add tag file parsing, better ELF parsing, basic output
Added a parser for the tag file. The tag file is expected to be
composed of:
[atom|ptr] <symbol> : <tag-name>\n

Each line describes one symbol.

The tag infromation and ELF information are combined to form an
output. The output contains lines of form:
<hex-address>,<size>,<tag-name>\n

In case that the variable to be tagged is a pointer, the address to
which the pointer points is also tagged (only the first byte). TODO
think about this more.

Better use of exceptions, better memory management.

Save string names of section headers together with them.