Add policy
Added construction of the policy from the AST. The policy consists of
topologies that are defined in a file. Two types of topology are
stored:
- basic: Holds the graph as an adjacency matrix with a mapping
  corresponding a tag to its row in the matrix.
- linear: Holds the graph as a vector (graph is a linear line).

First the basic and linear topologies are read, then the expression
topologies.

The expression topology is resolved into a basic topology. It
traverses the expression and creates it. Rules:
- +: For summation of two basic topologies, we use a direct sum on the
  adjacency matrixes.
- *: For product of two basic topologies, we use a cartesian product
  of graphs, which is calculated as (A (x) I_2) + (I_1 (x) B),
  where (x) represents Kronecker product.

Currently only topologies are supported.