Output a file, minor refactoring
The program outputs a file, consisting of two parts. The first part
contains the policy in the form of a matrix. The first line of the
file indicates the size of the matrix. Next 'n' lines contain the name
of the tag and its row.
Lines after that contain the information about tagged memory. Each
line contains the address, size of the memory and the tag.

Tag is currently defined only as a row index of the policy matrix.