Check policy for DAG, add mapping from index to tag name
The final topology is checked, if the graph is a DAG. This property is
necessary for tag propagation.

Added a mapping to the basic topology that links the row index of a
tag to its name. It is basically a reversed 'toindex', that maps from
the tag name to its index.

The product tags are now written as 'tag1,tag2', not 'tag1*tag2'.