Add two ways of computing LCA
Added two ways of computing LCA.

First is discarded and is located in 'lca.cc'. It computes the LCA
tree from the DAG, which would be used to calculate the LSCA. The
problem with LSCA is that we quickly reach the maximum tag, because
for LSCA we demand that all paths to nodes x and y go through the
result.

The second one computes all LCA pairs of a DAG by creating a
transitive closure of the graph and then by processing it. Currently
looks to work fine, but there can be multiple candidates for the
result - currently the node that is handled the last out of the
candidates will be the result.