Change handling of pointer tags
The pointer tag now expects a 'size' argument, which gives information
on the size of the object that the pointer points to. This allows us
to tag the memory of the object better.

The format of the pointer tag declaration is now of form:

ptr <name> size = <unsigned long> : <tag>\n