Change perimiter guards and CFG for policy
The policy specification now has strings, denoted as "string". The tag
information for basic and linear topologies are expected to be
strings, not identifiers anymore. This is done so that tags can have
special symbols, such as '(' in the names, without hurting the
grammar.

The perimiter guard is defined in the policy file as

pg <name> {
    <type>|<file> : <in/out/err>|<str>
	tag = str
}

It can be a 'type' - then we expect keywords 'in', 'out' or 'err', as
in standrad input / output / error. In case of a file we expect a
string containing a file name. The tag is expected to be a string.