Add execution traces for hello-policy test case
The 'hello-policy' is redesigned - it has 2 runs with two different
sets of tags specified for the program. Each run has an execution
trace and register / memory data attached to it. The spike-tag uses a
debugger to gather the execution trace and to read registers at
apropriate times (before ecalls).

Added new bash script for running the test case.